requirements.txt
praat-parselmouth
librosa
soundfile
scikit-learn
joblib
flask
pyngrok
ffmpeg-python
waitress